@@ -15,18 +15,18 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@babel/helper-create-class-features-plugin": "^7.26.10",
-    "@babel/preset-env": "^7.26.10",
+    "@babel/helper-create-class-features-plugin": "^7.28.0",
+    "@babel/preset-env": "^7.28.0",
     "@docusaurus/plugin-google-analytics": "^3.9.2",
     "@docusaurus/preset-classic": "^3.9.2",
-    "@mdx-js/react": "^3.1.0",
+    "@mdx-js/react": "^3.1.1",
     "classnames": "^2.5.1",
-    "clsx": "^1.1.1",
-    "docusaurus-plugin-internaldocs-fb": "1.19.1",
-    "loader-utils": "3.2.1",
-    "prism-react-renderer": "^1.3.3",
-    "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "clsx": "^2.0.0",
+    "docusaurus-plugin-internaldocs-fb": "1.19.2",
+    "loader-utils": "3.3.1",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "browserslist": {
     "production": [
@@ -46,6 +46,6 @@
     "yarn": "^1.5"
   },
   "devDependencies": {
-    "yarn-audit-fix": "^9.3.10"
+    "yarn-audit-fix": "^10.0.5"
   }
 }
