@@ -6,8 +6,7 @@
  */
 
 import {fbContent} from 'docusaurus-plugin-internaldocs-fb/internal';
-const lightCodeTheme = require('prism-react-renderer/themes/github');
-const darkCodeTheme = require('prism-react-renderer/themes/dracula');
+import {themes} from 'prism-react-renderer';
 
 // With JSDoc @type annotations, IDEs can provide config autocompletion
 /** @type {import('@docusaurus/types').DocusaurusConfig} */
@@ -77,8 +76,8 @@ const darkCodeTheme = require('prism-react-renderer/themes/dracula');
       },
       prism: {
         additionalLanguages: ['swift'],
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: themes.github,
+        darkTheme: themes.darcula,
       },
     }),
 });
